+++ Environment_oM/Environment_oM.csproj
@@ -48,6 +48,7 @@
     <Compile Include="Climate\Time.cs" />
     <Compile Include="Climate\Sun.cs" />
     <Compile Include="Elements\Building.cs" />
+    <Compile Include="Elements\BoundaryZone.cs" />
     <Compile Include="Elements\Mesh.cs" />
     <Compile Include="Elements\Edge.cs" />
     <Compile Include="Elements\Enums\BuildingType.cs" />
